updates styles of header buttons
