+++ src/gatsby-theme-docz/components/Header/index.js
@@ -1,3 +1,10 @@
-import React from "react";
+/** @jsx jsx */
+import { jsx } from "theme-ui";
+import * as styles from "./styles";
+import { HeaderButtons } from "../HeaderButtons";
 
-export const Header = () => <></>;
+export const Header = () => (
+  <div sx={styles.buttons}>
+    <HeaderButtons />
+  </div>
+);
