+++ src/gatsby-theme-docz/components/Header/styles.js
@@ -1,0 +1,6 @@
+export const buttons = {
+  position: "fixed",
+  top: 1,
+  right: 1,
+  zIndex: "999",
+};
