+++ src/gatsby-theme-docz/components/Sidebar/index.js
@@ -2,12 +2,11 @@
 /** @jsxFrag React.Fragment */
 import React, { useEffect, useRef, useState } from "react";
 import { Box, jsx } from "theme-ui";
-import { useConfig, useCurrentDoc, useMenus } from "docz";
+import { useCurrentDoc, useMenus } from "docz";
 import { NavSearch } from "gatsby-theme-docz/src/components/NavSearch";
 import { NavLink } from "gatsby-theme-docz/src/components/NavLink";
 import { NavGroup } from "gatsby-theme-docz/src/components/NavGroup";
 import { Logo } from "gatsby-theme-docz/src/components/Logo";
-import { Edit, Github } from "gatsby-theme-docz/src/components/Icons";
 import * as styles from "./styles";
 
 export const Sidebar = React.forwardRef((props, ref) => {
@@ -15,15 +14,8 @@ export const Sidebar = React.forwardRef((props, ref) => {
   const menus = useMenus({ query });
   const currentDoc = useCurrentDoc();
   const currentDocRef = useRef();
-  const {
-    repository,
-    themeConfig: { showMarkdownEditButton },
-  } = useConfig();
-  const { edit = true, ...doc } = useCurrentDoc();
+  const handleChange = ev => setQuery(ev.target.value);
 
-  const handleChange = ev => {
-    setQuery(ev.target.value);
-  };
   useEffect(() => {
     if (ref.current && currentDocRef.current) {
       ref.current.scrollTo(0, currentDocRef.current.offsetTop);
@@ -68,29 +60,6 @@ export const Sidebar = React.forwardRef((props, ref) => {
               </div>
             );
           })}
-        <div sx={styles.buttons}>
-          {repository && (
-            <a
-              href={repository}
-              sx={styles.button({ color: "var(--color-blue--dark)" })}
-              target="_blank"
-              rel="noopener noreferrer"
-            >
-              <Github size={15} />
-            </a>
-          )}
-
-          {showMarkdownEditButton && edit && doc.link && (
-            <a
-              sx={styles.button}
-              href={doc.link}
-              target="_blank"
-              rel="noopener noreferrer"
-            >
-              <Edit width={15} />
-            </a>
-          )}
-        </div>
       </Box>
     </>
   );
